requirements.txt
flask==2.2.2
requests==2.28.1
mysql-connector-python==8.0.30
